Fix bucket not updating lambda.zip file
As the `lambda.zip` file always has the same name,
the bucket did not know that the files inside the zip
were modified, then it did not update them because
it thought it was the same file.

Now, the lambda.zip file name also has the exaclty date and time
of the deploy, then it should update correctly now.